requirements.txt
-e svn+http://code.djangoproject.com/svn/django/trunk/#egg=Django
django-tinymce
south>=0.7.3
flup
-e git://github.com/pydanny/django-uni-form.git#egg=uni_form
bleach
django-debug-toolbar
pylibmc
mysql-python
scrapy
#python-cloudfiles
#PIL
#django-simple-captcha